Added delete condition for build folder

+++ install.sh
@@ -19,7 +19,7 @@ fi
 cd ..
 
 echo "Building TileRummy2D..."
-rm -r build
+[ -d build ] && rm -r build
 mkdir build
 cd build
 cmake -DCMAKE_BUILD_TYPE=Release ..
